+++ backend/src/main/java/com/warehouse/inventory/dto/request/UpdateProductRequest.java
@@ -14,7 +14,7 @@ public record UpdateProductRequest(
     String description,
 
     @NotNull(message = "Category ID is required.")
-    Long categoryId,
+    Integer categoryId,
 
     @NotBlank(message = "Unit of measure must not be blank.")
     @Size(max = 20, message = "Unit of measure must not exceed 20 characters.")
