+++ backend/src/main/java/com/warehouse/inventory/service/InventoryQuery.java
@@ -3,9 +3,9 @@ package com.warehouse.inventory.service;
 import com.warehouse.inventory.domain.enums.StockStatus;
 
 public record InventoryQuery(
-    Long locationId,
-    Long productId,
-    Long categoryId,
+    Integer locationId,
+    Integer productId,
+    Integer categoryId,
     StockStatus stockStatus,
     String search,
     String sortBy,
