+++ backend/src/main/java/com/warehouse/inventory/service/ProductQuery.java
@@ -2,7 +2,7 @@ package com.warehouse.inventory.service;
 
 public record ProductQuery(
     String search,
-    Long categoryId,
+    Integer categoryId,
     Boolean active,
     String sortBy,
     String sortDir,
